Avoid filesystem access to improve best ref perf
High I/O and accessing an incredible number of files adds overhead that
reduces the performance of Asklepian. This change attacks two problems:

* Accessing the QC file for each FASTA sequence to determine whether a
  given consensus reference is the best or not. To circumvent this, I
  have put a significant amount of work into refreshing the GET PAG API
  endpoints on Majora (see samstudio/majora#42). This allows Asklepian
  to request a list of all the latest, non-supressed, QC passing
  consensus sequences along with the QC metrics used for get best.

* Accessing the FASTA sequence for each best consensus. I have solved
  this by using the latest Elan matched consensus that contains all the
  sequences that have passed QC according to Elan. I iterate through the
  sequences and check whether the sequences' PAG name is in the best
  set. We write to stdout if it is, and skip if it is not.

This change prevents having to access the file system to determine QC
bestness or to fetch the consensus sequences. I have tested this locally
and determined it matches yesterday's output. I anticipate this change
to save between 60 and 90 minutes a day going forward.